feat/ref: 유효성 검사 추가, 비밀번호 검증 수정
- 비밀번호를 검증하는 로직을 repository -> service로 이동
- 유효성 검사 @Valid 추가, 요구사항 lv6 수행

+++ src/main/java/com/example/schedule/service/ScheduleServiceImpl.java
@@ -10,7 +10,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Map;
 
 @Service
 public class ScheduleServiceImpl implements ScheduleService {
@@ -35,58 +34,63 @@ public class ScheduleServiceImpl implements ScheduleService {
         scheduleRepository.createSchedule(schedule);
     }
 
-    // 전체 일정 조회
+    // 일정 조회 기능
+    // 전체 조회
     @Override
     public List<ScheduleResponseDto> findAllSchedules() {
         return scheduleRepository.findAllSchedules();
     }
 
-    // 일정 id
+    // 단건 조회: id
     @Override
     public ScheduleResponseDto findScheduleById(int id) {
         return scheduleRepository.findScheduleById(id);
     }
 
-    // 작성자
+    // 다건 조회: 작성자
     @Override
     public List<ScheduleResponseDto> findSchedulesByUserName(String userName) {
         return scheduleRepository.findSchedulesByUserName(userName);
     }
 
-    // 수정일
+    // 다건 조회: 수정일 기준 기간 조회
     @Override
     public List<ScheduleResponseDto> findSchedulesByUpdatedDate(String startDate, String endDate) {
         return scheduleRepository.findSchedulesByUpdatedDate(startDate, endDate);
     }
 
     // 일정 수정
+    @Transactional
     @Override
     public void updateSchedule(ScheduleUpdateRequestDto dto) {
-        Map<String, Object> credentials = scheduleRepository.checkPassword(dto.getId());
-        String storedPassword = (String) credentials.get("password");
-        int userId = ((Number) credentials.get("user_id")).intValue();
-
-        if (!storedPassword.equals(dto.getPassword())) {
+        Schedule schedule = scheduleRepository.findById(dto.getId());
+        if(schedule == null){
+            throw new IllegalArgumentException("존재하지 않는 일정");
+        }
+        if (!schedule.getPassword().equals(dto.getPassword())) {
             throw new IllegalArgumentException("비밀번호 불일치");
         }
-
         int updatedCount = scheduleRepository.updateScheduleTask(dto.getId(), dto.getUpdateTask());
         if (updatedCount <= 0) {
             throw new IllegalArgumentException("일정 수정 불가");
         }
-
-        int userUpdatedCount = userRepository.updateUserName(userId, dto.getUpdateUserName());
+        int userUpdatedCount = userRepository.updateUserName(schedule.getUserId(), dto.getUpdateUserName());
         if (userUpdatedCount <= 0) {
             throw new IllegalArgumentException("사용자 이름 수정 불가");
         }
     }
 
     // 일정 삭제
+    @Transactional
     @Override
     public void deleteSchedule(int scheduleId, String password) {
-        int deletedCount = scheduleRepository.deleteSchedule(scheduleId, password);
-        if (deletedCount <= 0) {
-            throw new IllegalArgumentException("비밀번호 불일치 or 존재하지 않는 일정");
+        Schedule schedule = scheduleRepository.findById(scheduleId);
+        if(schedule == null){
+            throw new IllegalArgumentException("존재하지 않는 일정");
         }
+        if(!schedule.getPassword().equals(password)){
+            throw new IllegalArgumentException("비밀번호 불일치");
+        }
+        scheduleRepository.deleteSchedule(scheduleId, password);
     }
 }
